+++ Task-Monitor/src/main/java/com/example/Task/Monitor/Domain/Service/TaskService.java
@@ -1,6 +1,7 @@
 package com.example.Task.Monitor.Domain.Service;
 
 import com.example.Task.Monitor.Domain.Entity.Task;
+import com.example.Task.Monitor.Exceptions.NoIdExistsException;
 import com.example.Task.Monitor.Repository.TaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,7 +15,7 @@ public class TaskService {
     public TaskRepository taskRepository;
 
     @Autowired
-    public TaskService (TaskRepository taskRepository) {
+    public TaskService(TaskRepository taskRepository) {
         this.taskRepository = taskRepository;
     }
 
@@ -22,9 +23,14 @@ public class TaskService {
         return taskRepository.findAll();
     }
 
-    public void createTask(Task task) {taskRepository.save(task);}
+    public void createTask(Task task) {
+        taskRepository.save(task);
+    }
 
     public void updateTask(Task task, UUID id) {
+        if (taskRepository.findByTaskId(id) == null) {
+            throw new NoIdExistsException();
+        }
         Task myTask = taskRepository.findByTaskId(id);
         myTask.setTitle(task.getTitle());
         myTask.setDescription(task.getDescription());
@@ -35,6 +41,9 @@ public class TaskService {
     }
 
     public void deleteTask(UUID id) {
+        if (taskRepository.findByTaskId(id) == null) {
+            throw new NoIdExistsException();
+        }
         taskRepository.deleteByTaskId(id);
     }
 
@@ -42,10 +51,16 @@ public class TaskService {
         return taskRepository.findFiveBestEmployeeId();
     }
 
-    public List<UUID> topThreeClientId() { return taskRepository.findThreeTopClientId();}
+    public List<UUID> topThreeClientId() {
+        return taskRepository.findThreeTopClientId();
+    }
 
-    public List<UUID> worstThreeClientId() { return taskRepository.findThreeWorstClientId();}
+    public List<UUID> worstThreeClientId() {
+        return taskRepository.findThreeWorstClientId();
+    }
 
-    public Integer countClients(UUID id) {return taskRepository.countByClient(id);}
+    public Integer countClients(UUID id) {
+        return taskRepository.countByClient(id);
+    }
 
 }
